fix: fix claims and token data response

+++ src/common/jwt.rs
@@ -5,5 +5,6 @@ use uuid::Uuid;
 pub struct Claims {
     pub user_id: Uuid,
     pub stripe_customer_id: String,
-    pub exp: usize,
+    pub exp: u32,
 }
+
